+++ pkg/converter/converter_test.go
@@ -21,6 +21,7 @@ import (
 	"github.com/dynatrace/dynatrace-configuration-as-code/internal/template"
 	config "github.com/dynatrace/dynatrace-configuration-as-code/pkg/config/v2"
 	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/config/v2/parameter"
+	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/converter/v1environment"
 	"reflect"
 	"testing"
 
@@ -1114,4 +1115,63 @@ func Test_convertReservedParameters(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestNewEnvironmentDefinitionFromV1(t *testing.T) {
+	type args struct {
+		env   *v1environment.EnvironmentV1
+		group string
+	}
+	tests := []struct {
+		name string
+		args args
+		want manifest.EnvironmentDefinition
+	}{
+		{
+			"simple v1 environment is converted",
+			args{
+				v1environment.NewEnvironmentV1("test", "name", "group", "http://google.com", "NAME"),
+				"group",
+			},
+			createValueEnvironmentDefinition(),
+		},
+		{
+			"v1 environment with env var is converted",
+			args{
+				v1environment.NewEnvironmentV1("test", "name", "group", "{{ .Env.ENV_VAR }}", "NAME"),
+				"group",
+			},
+			createEnvEnvironmentDefinition(),
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := manifest.NewEnvironmentDefinition(tt.args.env.GetId(), newUrlDefinitionFromV1(tt.args.env), tt.args.group, manifest.Token{Name: tt.args.env.GetTokenName()}); !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("NewEnvironmentDefinitionFromV1() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+func createEnvEnvironmentDefinition() manifest.EnvironmentDefinition {
+	return manifest.NewEnvironmentDefinition(
+		"test",
+		manifest.UrlDefinition{
+			Type: manifest.EnvironmentUrlType,
+			Name: "ENV_VAR",
+		},
+		"group",
+		manifest.Token{Name: "NAME"},
+	)
+}
+
+func createValueEnvironmentDefinition() manifest.EnvironmentDefinition {
+	return manifest.NewEnvironmentDefinition(
+		"test",
+		manifest.UrlDefinition{
+			Type:  manifest.ValueUrlType,
+			Value: "http://google.com",
+		},
+		"group",
+		manifest.Token{Name: "NAME"},
+	)
 }
